requirements.txt
pandas==1.5.3
serpapi==0.1.5
python-dotenv==0.21.0
scikit-learn==1.1.2
sqlalchemy==2.0.10
numpy==1.23.4
google-search-results==2.4.2
psycopg2-binary==2.9.3 
pyyaml==6.0.2
pydantic==2.11.2